Fix: fix bug with restricted JSON response.

+++ src/main/kotlin/com/example/topsaleparser/domain/Product.kt
@@ -1,12 +1,10 @@
 package com.example.topsaleparser.domain
 
-import java.math.BigDecimal
-
 data class Product(
-    val id: String? = null,
-    val title: String,
-    val currentPrice: BigDecimal,
-    val oldPrice: BigDecimal,
-    val imgLink: String? = null,
-    val productLink: String? = null
+        val id: String? = null,
+        val title: String,
+        val currentPrice: Int,
+        val oldPrice: Int,
+        val imgLink: String? = null,
+        val productLink: String? = null
 )
